review: string+ extension에 있는 함수를 DTO에서 실행하도록 수정
