+++ Koin/Core/Extensions/String+.swift
@@ -26,88 +26,6 @@ extension String {
         return dateFormatter.date(from: self)
     }
 
-    func modifyFontInHtml() -> String? {
-        do {
-            let document = try SwiftSoup.parse(self)
-            try document.select("br").remove()
-            
-            let paragraphs = try document.select("p")
-            let bTags = try document.select("b")
-            let iTags = try document.select("i")
-            let hTags = try document.select("h1, h2, h3, h4, h5, h6")
-            let spans = try document.select("span")
-            let ulTags = try document.select("ul")
-            let olTags = try document.select("ol")
-            let liTags = try document.select("li")
-            
-            var emptyPCount = 0
-            for paragraph in paragraphs {
-                let paragraphText = try paragraph.text()
-                if paragraphText.isEmpty {
-                    emptyPCount += 1
-                } else {
-                    emptyPCount = 0
-                }
-                
-                if emptyPCount > 1 {
-                    try paragraph.remove()
-                }
-                
-                if (try paragraph.select("img").first()) != nil {
-                    try paragraph.append("<p></p>")
-                }
-            }
-        
-            let allTags: [Elements] = [paragraphs, bTags, iTags, hTags, spans, ulTags, olTags, liTags]
-            for elements in allTags {
-                for element in elements {
-                    let existingStyle = try element.attr("style")
-                    let newStyle = "font-family: 'Pretandard-Medium', sans-serif; font-size: 16px;"
-                    let updatedStyle = existingStyle.isEmpty ? newStyle : "\(existingStyle); \(newStyle)"
-                    try element.attr("style", updatedStyle)
-                }
-            }
-           
-            for element in bTags {
-                let existingStyle = try element.attr("style")
-                let newBoldStyle = "font-family: 'Pretandard-Bold', sans-serif; font-size: 16px;"
-                let updatedBoldStyle = existingStyle.isEmpty ? newBoldStyle : "\(existingStyle); \(newBoldStyle)"
-                try element.attr("style", updatedBoldStyle)
-            }
-            
-            let setHeightUsingCSS = """
-                <head>
-                    <style type="text/css">
-                        img {
-                            max-height: 100%;
-                            max-width: 317px !important;
-                            width: auto;
-                            height: auto;
-                            display: block;
-                            margin-bottom: 20px; /* Space below the image */
-                        }
-                        .txc-image {
-                            text-align: center; /* Center the image */
-                        }
-                        .bc-s-post-ctnt-area {
-                            padding: 20px; /* Optional padding */
-                            border-top: 1px solid transparent; /* Prevent margin collapse */
-                        }
-                    </style>
-                </head>
-                <body class="center">
-                    <div class="bc-s-post-ctnt-area">
-                        \(try document.body()?.html() ?? "")
-                    </div>
-                </body>
-                """
-            return setHeightUsingCSS
-        } catch {
-            print("Error processing HTML: \(error)")
-            return nil
-        }
-    }
-    
     func extractImageStringFromHtmlTag() -> String? {
         do {
             let docs: Document = try SwiftSoup.parse(self)
